Add additional suppressions for backwards compatibility.
Many of our projects still have functions like `public AppController:_auth()`, which break the "normal" rules. These could be changed to comply more strictly with the code standards, but in cases like this the underscore helps emphasize that the controller method isn't meant to be handled by a URL route.