WIP (Lots of stuff)
New Sniff is non-functional in its current state. Needs review and repair.

The README has notes on improving the testing of our coding standard itself, to ensure we're catching all of the rules we want to be catching and that none of the intentionally-failing tests are failing for the wrong infraction. One thought is to somehow scan all of the errors/warnings from our set of sniffs (like we scan tokens in the CakePHP Skeleton's token replacement) and make sure the code sniffer output during the test run contains every error or warning name we expect.

We also need to try to make sure the code in the bundled sniffs (at least) complies as closely as possible with our own ruleset.