Test how much of the code standard is covered.
Compares the annotations in the sample files intended to fail against the total list of sniffs the Standard includes and emits an assertion failure for "not covered" sniffs. 

The test is currently disabled, because we're only testing 32 of the 112 sniffs.

Round out test coverage now that some sniffs are not being suppressed thanks to the name collision on the `tests/` folder.